fix crash on first startup with the new data cleaner

+++ app/src/main/java/de/tadris/fitness/util/DataManager.java
@@ -32,6 +32,9 @@ public class DataManager {
 
     public static void cleanFiles(Context context) {
         File dir = new File(getSharedDirectory(context));
+        if (!dir.exists()) {
+            return;
+        }
         for (File file : dir.listFiles()) {
             if (file.isFile()) {
                 if (file.delete()) {
